# high-load-systems/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CDN Control Plane API",
    "description": "\nControl plane for a content delivery network: the service that decides **where**\ncontent lives and **which** edge node a client should talk to. It never serves the\ncontent bytes itself.\n\n* **Node management** — registration, lifecycle and continuous telemetry ingest.\n* **Routing** — ranked edge-node selection by proximity, load and cache state.\n* **Distribution** — asset metadata and the regional policy governing where it may be cached.\n* **Invalidation** — tracked purge campaigns with per-node acknowledgement.\n\nErrors follow RFC 9457 and are returned as `application/problem+json`.\nEvery response carries `X-Instance-ID`, `X-Request-ID` and `X-Response-Time-ms`.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/nodes": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "List edge nodes",
        "operationId": "list_nodes_api_v1_nodes_get",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by region code.",
              "title": "Region"
            },
            "description": "Filter by region code."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NodeStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_NodeRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "nodes"
        ],
        "summary": "Register an edge node",
        "operationId": "register_node_api_v1_nodes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeRegister"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state of the resource.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nodes/{node_id}": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "Get one edge node",
        "operationId": "get_node_api_v1_nodes__node_id__get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "nodes"
        ],
        "summary": "Update node status or capacity",
        "operationId": "update_node_api_v1_nodes__node_id__patch",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NodeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NodeRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state of the resource.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation or violated a business rule.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "nodes"
        ],
        "summary": "Decommission an edge node",
        "operationId": "delete_node_api_v1_nodes__node_id__delete",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nodes/{node_id}/heartbeat": {
      "post": {
        "tags": [
          "nodes"
        ],
        "summary": "Report node telemetry",
        "description": "Ingest one telemetry sample.\n\nAnswers 202 rather than 200: the sample is recorded, but the fleet-wide\nconsequences of it — routing weights, purge targeting — are observed by other\nendpoints, not computed and returned here.",
        "operationId": "report_heartbeat_api_v1_nodes__node_id__heartbeat_post",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatReport"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatAccepted"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nodes/{node_id}/replicas": {
      "get": {
        "tags": [
          "nodes"
        ],
        "summary": "List content cached on a node",
        "operationId": "list_replicas_api_v1_nodes__node_id__replicas_get",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReplicaRead"
                  },
                  "title": "Response List Replicas Api V1 Nodes  Node Id  Replicas Get"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/nodes/{node_id}/replicas/{asset_id}": {
      "put": {
        "tags": [
          "nodes"
        ],
        "summary": "Report cache state for one asset",
        "description": "Declare what this node holds for one asset.\n\nPUT, because an agent reports the full current state of one replica and repeating\nthe same report must not change the outcome.",
        "operationId": "report_replica_api_v1_nodes__node_id__replicas__asset_id__put",
        "parameters": [
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Node Id"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Asset Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplicaReport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplicaRead"
                }
              }
            }
          },
          "404": {
            "description": "Referenced resource does not exist.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Request failed validation or violated a business rule.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeartbeatAccepted": {
        "properties": {
          "node_id": {
            "type": "string",
            "format": "uuid",
            "title": "Node Id"
          },
          "status": {
            "$ref": "#/components/schemas/NodeStatus"
          },
          "current_load_percent": {
            "type": "integer",
            "title": "Current Load Percent"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At"
          },
          "next_heartbeat_in_seconds": {
            "type": "integer",
            "title": "Next Heartbeat In Seconds"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "status",
          "current_load_percent",
          "received_at",
          "next_heartbeat_in_seconds"
        ],
        "title": "HeartbeatAccepted"
      },
      "HeartbeatReport": {
        "properties": {
          "reported_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reported At"
          },
          "cpu_percent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cpu Percent"
          },
          "memory_percent": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Memory Percent"
          },
          "bandwidth_out_mbps": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Bandwidth Out Mbps"
          },
          "active_connections": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Active Connections"
          },
          "cache_hit_ratio": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Hit Ratio"
          }
        },
        "type": "object",
        "required": [
          "cpu_percent",
          "memory_percent",
          "bandwidth_out_mbps",
          "active_connections"
        ],
        "title": "HeartbeatReport",
        "description": "Telemetry sample sent by an edge agent.\n\n`reported_at` is optional and defaults to server receipt time; an agent with a\nskewed clock should not be able to poison the freshness calculation, so routing\neligibility is evaluated against `received_at`, never this field."
      },
...
      "NodeRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "hostname": {
            "type": "string",
            "title": "Hostname"
          },
          "public_ipv4": {
            "type": "string",
            "format": "ipv4",
            "title": "Public Ipv4"
          },
          "region_code": {
            "type": "string",
            "title": "Region Code"
          },
          "status": {
            "$ref": "#/components/schemas/NodeStatus"
          },
          "capacity_mbps": {
            "type": "integer",
            "title": "Capacity Mbps"
          },
          "agent_version": {
            "type": "string",
            "title": "Agent Version"
          },
          "last_heartbeat_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          },
          "current_load_percent": {
            "type": "integer",
            "title": "Current Load Percent"
          },
          "active_connections": {
            "type": "integer",
            "title": "Active Connections"
          },
          "cache_hit_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Hit Ratio"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "hostname",
          "public_ipv4",
          "region_code",
          "status",
          "capacity_mbps",
          "agent_version",
          "last_heartbeat_at",
          "current_load_percent",
          "active_connections",
          "cache_hit_ratio",
          "created_at"
        ],
        "title": "NodeRead"
      },
      "NodeRegister": {
        "properties": {
          "hostname": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Hostname",
            "examples": [
              "edge-eu-01.cdn.net"
            ]
          },
          "public_ipv4": {
            "type": "string",
            "format": "ipv4",
            "title": "Public Ipv4"
          },
          "region_code": {
            "type": "string",
            "maxLength": 32,
            "minLength": 2,
            "pattern": "^[a-z0-9]+(-[a-z0-9]+)*$",
            "title": "Region Code",
            "examples": [
              "eu-central"
            ]
          },
          "capacity_mbps": {
            "type": "integer",
            "maximum": 1000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Capacity Mbps",
            "examples": [
              10000
            ]
          },
          "agent_version": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Agent Version",
            "examples": [
              "1.4.2"
            ]
          }
        },
        "type": "object",
        "required": [
          "hostname",
          "public_ipv4",
          "region_code",
          "capacity_mbps",
          "agent_version"
        ],
        "title": "NodeRegister"
      },
      "NodeStatus": {
        "type": "string",
        "enum": [
          "provisioning",
          "healthy",
          "degraded",
          "draining",
          "offline"
        ],
        "title": "NodeStatus"
      },
      "NodeUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/NodeStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "capacity_mbps": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Mbps"
          },
          "agent_version": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Version"
          }
        },
        "type": "object",
        "title": "NodeUpdate"
      },
...
      "Page_NodeRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NodeRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[NodeRead]"
      },
...
      "ReplicaRead": {
        "properties": {
          "node_id": {
            "type": "string",
            "format": "uuid",
            "title": "Node Id"
          },
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Asset Id"
          },
          "state": {
            "$ref": "#/components/schemas/ReplicaState"
          },
          "cached_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Version"
          },
          "bytes_cached": {
            "type": "integer",
            "title": "Bytes Cached"
          },
          "last_verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Verified At"
          }
        },
        "type": "object",
        "required": [
          "node_id",
          "asset_id",
          "state",
          "cached_version",
          "bytes_cached",
          "last_verified_at"
        ],
        "title": "ReplicaRead"
      },
      "ReplicaReport": {
        "properties": {
          "state": {
            "$ref": "#/components/schemas/ReplicaState"
          },
          "cached_version": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Version"
          },
          "bytes_cached": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Bytes Cached",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "ReplicaReport",
        "description": "An edge agent declaring what it currently holds for one asset."
      },
      "ReplicaState": {
        "type": "string",
        "enum": [
          "pending",
          "cached",
          "stale",
          "purged"
        ],
        "title": "ReplicaState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Problem": {
        "title": "Problem",
        "description": "RFC 9457 problem details.",
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ]
      }
    }
  }
}